Moved StatsInteractor connectors to different files
