+++ Sources/FilterState/Proxies/GroupAccessor.swift
@@ -30,8 +30,6 @@ extension FiltersContainer {
     return .init(filtersContainer: self, groupName: groupName)
   }
   
-  
-  
 }
 
 public class ReadOnlyFiltersContainer {
@@ -53,8 +51,6 @@ public class ReadOnlyFiltersContainer {
   public subscript(hierarchical groupName: String) -> ReadOnlyGroupAccessor<Filter.Facet> {
     return ReadOnlyGroupAccessor(filtersContainer[hierarchical: groupName])
   }
-  
-  
   
 }
